Struct pallet_session_validator_management::pallet::Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn set( origin: OriginFor<T>, validators: BoundedVec<T::CommitteeMember, T::MaxValidators>, for_epoch_number: T::ScEpochNumber, selection_inputs_hash: SizedByteString<32>, ) -> DispatchResult

โfor_epoch_numberโ parameter is needed only for validation purposes, because we need to make sure that check_inherent uses the same epoch_number as was used to create inherent data. Alternative approach would be to put epoch number inside InherentData. However, sidechain epoch number is set in Runtime, thus, inherent data provider doesnโt have to know about it. On top of that, the latter approach is slightly more complicated to code.

pub fn set_main_chain_scripts( origin: OriginFor<T>, committee_candidate_address: MainchainAddress, d_parameter_policy_id: PolicyId, permissioned_candidates_policy_id: PolicyId, ) -> DispatchResult

Changes the main chain scripts used for committee rotation.

This extrinsic must be run either using sudo or some other chain governance mechanism.

impl<T: Config> Pallet<T>

pub fn get_next_unset_epoch_number() -> T::ScEpochNumber

pub fn get_current_authority(index: usize) -> Option<T::AuthorityId>

pub fn get_current_authority_round_robin( index: usize, ) -> Option<T::CommitteeMember>

pub fn current_committee_storage() -> CommitteeInfo<T::ScEpochNumber, T::CommitteeMember, T::MaxValidators>

pub fn next_committee_storage() -> Option<CommitteeInfo<T::ScEpochNumber, T::CommitteeMember, T::MaxValidators>>

pub fn next_committee() -> Option<BoundedVec<T::AuthorityId, T::MaxValidators>>

This functionโs result should be always defined after inherent call of 1st block of each epoch

pub fn calculate_committee( authority_selection_inputs: T::AuthoritySelectionInputs, sidechain_epoch: T::ScEpochNumber, ) -> Option<Vec<T::CommitteeMember>>

pub fn rotate_committee_to_next_epoch() -> Option<Vec<T::CommitteeMember>>

pub fn get_current_committee() -> (T::ScEpochNumber, Vec<T::CommitteeMember>)

pub fn get_next_committee() -> Option<(T::ScEpochNumber, Vec<T::CommitteeMember>)>

pub fn get_main_chain_scripts() -> MainChainScripts