Crate pallet_session_validator_management

Pallet for setting the Partner Chain validators using inherent data

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

migrations

Implements storage migrations of pallet-session-validator-management.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

pallet_session_support

[pallet_session] integration. Implements Substrate’s [pallet_session].

session_manager

Implements pallet_partner_chains_session::SessionManager and pallet_partner_chains_session::ShouldEndSession, Partner Chain’s version of Substrate’s [pallet_session].

weights

Autogenerated weights for pallet_session_validator_management

Traits

CommitteeMember

Signifies that a type represents a committee member