pallet_session_validator_management::pallet

Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type MaxValidators: Get<u32>;
    type AuthorityId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen + Ord + Into<Self::AccountId>;
    type AuthorityKeys: Parameter + Member + MaybeSerializeDeserialize + Ord + MaxEncodedLen;
    type AuthoritySelectionInputs: Parameter;
    type ScEpochNumber: Parameter + MaxEncodedLen + Zero + Display + Add + One + Default + Ord + Copy + From<u64> + Into<u64>;
    type CommitteeMember: Parameter + Member + MaybeSerializeDeserialize + MaxEncodedLen + CommitteeMember<AuthorityId = Self::AuthorityId, AuthorityKeys = Self::AuthorityKeys>;
    type MainChainScriptsOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;

    // Required methods
    fn select_authorities(
        input: Self::AuthoritySelectionInputs,
        sidechain_epoch: Self::ScEpochNumber,
    ) -> Option<BoundedVec<Self::CommitteeMember, Self::MaxValidators>>;
    fn current_epoch_number() -> Self::ScEpochNumber;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type MaxValidators: Get<u32>

type AuthorityId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen + Ord + Into<Self::AccountId>

type AuthorityKeys: Parameter + Member + MaybeSerializeDeserialize + Ord + MaxEncodedLen

type AuthoritySelectionInputs: Parameter

type ScEpochNumber: Parameter + MaxEncodedLen + Zero + Display + Add + One + Default + Ord + Copy + From<u64> + Into<u64>

type CommitteeMember: Parameter + Member + MaybeSerializeDeserialize + MaxEncodedLen + CommitteeMember<AuthorityId = Self::AuthorityId, AuthorityKeys = Self::AuthorityKeys>

type MainChainScriptsOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin for governance calls

type WeightInfo: WeightInfo

Weight functions needed for pallet_session_validator_management.

Required Methods

fn select_authorities( input: Self::AuthoritySelectionInputs, sidechain_epoch: Self::ScEpochNumber, ) -> Option<BoundedVec<Self::CommitteeMember, Self::MaxValidators>>

fn current_epoch_number() -> Self::ScEpochNumber

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.